Add files via upload
Arrays are used to store multiple values in a single variable. Based on the dimensions and the type of data they store, arrays can be categorized into the following types:

1. One-Dimensional Array (1D Array):
It stores data in a single row or a single column (linear form). Example: int arr[5] = {1, 2, 3, 4, 5};


2. Two-Dimensional Array (2D Array):
It stores data in the form of a matrix (rows and columns). Example: int arr[3][3];


3. Multi-Dimensional Array:
An array with more than two dimensions. It is used for complex data structures like 3D matrices. Example: int arr[3][3][3];


4. Jagged Array (in some languages like Java or C#):
An array of arrays where each sub-array can have a different size.
5. Dynamic Array:
The size of the array can be changed during runtime (e.g., using ArrayList in Java or vector in C++).